Delete the "reduers/game/utils" module
Preparation to integrate the reducers' public API into the "reducers/index".
